fixed several remaining size_t issues

+++ libhdt/include/Iterator.hpp
@@ -261,7 +261,7 @@ public:
 	virtual ResultEstimationType numResultEstimation() {
 		return UNKNOWN;
 	}
-	virtual void skip(unsigned int /*pos*/) {
+	virtual void skip(size_t /*pos*/) {
 	/* Relative repositioning of index: ie. skip index by given argument places */
 	}
 };
@@ -274,7 +274,7 @@ public:
 		return false;
 	}
 
-	virtual unsigned int next() {
+	virtual size_t next() {
 		return 0;
 	}
 
@@ -289,10 +289,10 @@ class SequentialIteratorUInt: public IteratorUInt {
 private:
 	IteratorUInt* it1;
 	IteratorUInt* it2;
-	unsigned int offset; //the offset to apply to it2
+	size_t offset; //the offset to apply to it2
 
 public:
-	SequentialIteratorUInt(IteratorUInt* iterator1, IteratorUInt* iterator2, unsigned int offsetIt2):it1(iterator1), it2(iterator2), offset(offsetIt2){
+	SequentialIteratorUInt(IteratorUInt* iterator1, IteratorUInt* iterator2, size_t offsetIt2):it1(iterator1), it2(iterator2), offset(offsetIt2){
 	}
 	virtual ~SequentialIteratorUInt() { }
 
@@ -300,7 +300,7 @@ public:
 		return (it1->hasNext() || it2->hasNext());
 	}
 
-	unsigned int next() {
+	size_t next() {
 		if (it1->hasNext()){
 			return it1->next();
 		}
